niduc-projekt/requirements.txt
numpy==1.12.1
Pillow==4.1.1
